// src/types.d.ts
import { type TODO_FILTERS } from './conts'

export interface Todo {
  id: string
  title: string
  completed: boolean
}

export type TodoId = Pick<Todo, 'id'>
export type TodoTitle = Pick<Todo, 'title'>
export type TodoCompleted = Pick<Todo, 'completed'>

export type ListOfTodos = Todo[]
export type FiltersValue = typeof TODO_FILTERS[keyof typeof S_BUTTONS]